GraphQL from facebook
GraphQL is a data query language and runtime designed and used at Facebook to request and deliver data to mobile and web apps since 2012.

When we built Facebook's mobile applications, we needed a data-fetching API powerful enough to describe all of Facebook, yet simple enough to be easy to learn and use by our product developers. We developed GraphQL in 2012 to fill this need. Today it is the primary way we build client apps and the servers that drive them.

We've begun the process of open-sourcing GraphQL by drafting a specification, releasing a reference implementation, building tools, and forming a community around it here at graphql.org.
The reference implementation, graphql-js provides a library in JavaScript that is the basis for full GraphQL implementations and tools. It is not a fully standalone GraphQL server that a client developer could use to start manipulating and querying data. Most importantly it provides no mapping to a functioning, production-ready backend. For those using a express server, we recommend using express-graphql to provide a GraphQL server